coding_challenge/requirements.txt
asgiref==3.7.2
croniter==2.0.1
Django==4.2.8
djangorestframework==3.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
